% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{AbsoluteLoss}
\alias{AbsoluteLoss}
\title{Absolute values of residuals}
\usage{
AbsoluteLoss(r)
}
\arguments{
\item{r}{Residuals}
}
\value{
Absolute valued residuals
}
\description{
Absolute values of residuals
}

