% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr4pl-calculate.R
\name{calculate.dr4pl}
\alias{calculate.dr4pl}
\title{dr4pl-calculate}
\usage{
\method{calculate}{dr4pl}(x, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{x}{an object of class `dr4pl`}

\item{parm}{parameters of the dr4pl object. Usually made with [dr4pl_theta]}

\item{level}{confidence level to calculate. Defaults to 0.95}

\item{...}{extra arguments to be passed to [vcov.dr4pl]}
}
\description{
calculate various useful statistics.
}
