% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{coef.DR4PL}
\alias{coef.DR4PL}
\title{coef}
\usage{
\method{coef}{DR4PL}(object, ...)
}
\arguments{
\item{object}{A 'DR4PL' object}

\item{...}{arguments passed to coef}
}
\value{
A vector of parameters
}
\description{
Coefficient of a `DR4PL' object
}
