% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{coef.drra}
\alias{coef.drra}
\title{coef}
\usage{
\method{coef}{drra}(object, ...)
}
\arguments{
\item{object}{A 'drra' object}

\item{...}{arguments passed to coef}
}
\value{
A vector of parameters
}
\description{
Coefficient of a `drra' object
}

