% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{DerivativeF}
\alias{DerivativeF}
\title{Compute the derivative values of the mean response function.}
\usage{
DerivativeF(theta, x)
}
\arguments{
\item{theta}{Parameters of the 4PL model}

\item{x}{Dose levels}
}
\value{
Derivative values of the mean response function.
}
\description{
Compute the derivative values of the mean response function.
}
