% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr4pl-augment.R
\name{dr4pl-augment}
\alias{dr4pl-augment}
\alias{augment.dr4pl}
\title{Augment data with dr4pl}
\usage{
\method{augment}{dr4pl}(x, data = NULL, ...)
}
\arguments{
\item{x}{dr4pl object}

\item{data}{new data to use}

\item{...}{For future extension. Should not be used.}
}
\description{
Augment data with dr4pl
}
