% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{drc_error_1}
\alias{drc_error_1}
\title{Single High Outlier}
\description{
These are a handful of experimentally derived datasets from the wet-laboratory.
These all have numerical errors in other dose-response curve-packages, but not using these methods.
This data set exemplifies the case of a single extreme outlier of one dose measurement.
}
\keyword{sample_data}
