% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{drc_error_2}
\alias{drc_error_2}
\title{Multiple High Outliers at Different measurements}
\description{
These are a handful of experimentally derived datasets from the wet-laboratory.
These all have numerical errors in other dose-response curve-packages, but not using these methods.
This data set exemplifies the case of multiple outliers as well as a small number of observations per dose measurement.
}
\keyword{sample_data}
