% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{drc_error_4}
\alias{drc_error_4}
\title{Support Problem}
\description{
These are a handful of experimentally derived datasets from the wet-laboratory.
These all have numerical errors in other dose-response curve-packages, but not using these methods.
This data set exemplifies the support problem.
}
\keyword{sample_data}
