% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{drra.formula}
\alias{drra.formula}
\title{Fit a 4 parameter logistic (4PL) model to dose-response data.}
\usage{
\method{drra}{formula}(formula, data = list(), grad = NULL,
  init.parm = NULL, method.init = "logistic", method.optim = if
  (is.null(grad)) "Nelder-Mead" else "BFGS", method.robust = NULL, ...)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit. Either of the
form 'response ~ dose' or as a data frame with response values in first
column and dose values in second column.}

\item{data}{A data frame containing variables in the model.}

\item{grad}{A function returning the gradient values for the optimization
methods "BFGS", "CG" and "L-BFGS-B". If it is NULL, the Nelder-Mead method
will be applied.}

\item{init.parm}{A vector of initial parameters to be optimized in the model.}

\item{method.init}{The method of obtaining initial values of the parameters.
If it is NULL, a default "logistic" regression method will be used.}

\item{method.optim}{The method of optimization of the parameters. This method
name is directly applied to the \code{constrOptim} function provided in the
"base" package of R.}

\item{method.robust}{The robust estimation method to be used to fit a model.
If it is NULL, then the default least sum of squares estimator is used.}

\item{...}{Further arguments to be passed to \code{constrOptim}.}
}
\value{
A 'drra' object for which "confint", "gof", "print" and "summary"
  methods are implemented. For details, see the help page of each method.
  For example, type \code{?confint.drra} to obtain the confidence intervals
  of parameters of the 'drra' object.
}
\description{
A general 4PL model fitting function for analysis of
  dose-response relation.
}
\details{
This function fits a 4 parameter logistic (4PL) model to dose-response
  data. A formula of the model is
  \deqn{\theta[1]+(\theta[4]-\theta[1])/(1+(x/\theta[2])^\theta[3])}

  \code{method.init} specifies an initialization method to get initial parameter
  estimates based on data. The currently supported initialization methods are
  "Logistic", "Mead", "Anke". Details of the methods are given in Section
  "Details".

  \code{method.optim} specifies an optimization method to be used in
  "constrOptim" function. The currently supported optimization techniques
  include "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN" and "Brent". For
  further details, see the help page of \code{\link[stats]{optim}}.

  \code{method.robust} chooses a robust estimation method among 4 methods.
  The method of estimation is usually identified by the loss function of the
  method. This package implements 4 loss functions: sum of squares loss,
  absolute deviation loss, Huber's loss and Tukey's biweight loss. Each of
  loss function is explained in detail in the vignette.
}
\examples{
ryegrass.drra <- drra(rootl ~ conc, data = ryegrass)

ryegrass.drra
}
\author{
Hyowon An, Dirk P. Dittmer and J. S. Marron
}
\seealso{
\code{\link{confint.drra}}, \code{\link{gof.drra}},
\code{\link{print.drra}}, \code{\link{summary.drra}}
}

