% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{drra}
\alias{drra}
\alias{drra-package}
\alias{drra.default}
\alias{drraEst}
\title{Dose response relation analysis (drra)}
\usage{
drraEst(dose, response, constrained = constrained, grad, init.parm,
  method.init, method.optim, method.robust)

drra(x, ...)

\method{drra}{default}(dose, response, constrained = FALSE, grad = NULL,
  init.parm = NULL, method.init = "logistic", method.optim = if
  (is.null(grad)) "Nelder-Mead" else "BFGS", method.robust = NULL, ...)
}
\arguments{
\item{dose}{Dose}

\item{response}{Response}

\item{grad}{Gradient function}

\item{init.parm}{Vector of initial parameters}

\item{method.init}{Initialization method}

\item{method.optim}{Optimization method}

\item{method.robust}{Robust estimation method
- NULL: Sum of squares loss
- absolute: Absolute deviation loss
- Huber: Huber's loss
- Tukey: Tukey's biweight loss}

\item{x}{Object to drra.}

\item{...}{arguments passed to coef}
}
\description{
Main functions related to fitting

Fit the 4 parameter logistic model to the data using the function `drraEst'
}
\details{
Last updated: 09/19/2016
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Used in the default case, supplying a single dose and response variable
}}
\author{
Hyowon An, Lineberger Comprehensive Cancer Center
}

