% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{ErrFcn}
\alias{ErrFcn}
\title{Returns an error function for given robust fitting method}
\usage{
ErrFcn(method.robust)
}
\arguments{
\item{method.robust}{NULL, absolute, Huber, or Tukey
- NULL: Sum of squares loss
- absolute: Absolute deviation loss
- Huber: Huber's loss
- Tukey: Tukey's biweight loss}
}
\value{
Value of the sum of squared residuals
}
\description{
Returns an error function for given robust fitting method
}
