% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{gof.dr4pl}
\alias{gof.dr4pl}
\title{Perform the goodness-of-fit (gof) test for the 4PL model.}
\usage{
gof.dr4pl(object)
}
\arguments{
\item{object}{An object of the dr4pl class.}
}
\value{
Object of class `gof.dr4pl' which consists of .
}
\description{
Perform the goodness-of-fit (gof) test for the 4PL model when there
  are at least two replicates for each dose level.
}
\details{
Perform the goodness-of-fit (gof) test for the 4PL model in which the
  mean response actually follws the 4 Parameter Logistic model. There should
  be at least two replicates at each dose level. The test statistic follows the
  Chi squared distributions with the (n - 4) degrees of freedom where n is the
  number of observations and 4 is the number of parameters in the 4PL model. For
  detailed explanation of the method, please refer to Subsection 2.1.5 of
  Seber and Wild (1989).
}
\references{
\insertRef{Seber1989}{dr4pl}
}
