% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{gof}
\alias{gof}
\title{Perform the goodness-of-fit (gof) test for a model.}
\usage{
gof(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{dots for future extensions}
}
\value{
data.frame with goodness-of-fit results
}
\description{
S3 method for a model object
}
