% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{GradientSquaredLoss}
\alias{GradientSquaredLoss}
\title{Compute gradient values of the sum-of-squares loss function.}
\usage{
GradientSquaredLoss(theta, x, y)
}
\arguments{
\item{theta}{Parameters}

\item{x}{Dose}

\item{y}{Response}
}
\value{
Gradient values of the sum-of-squares loss function.
}
\description{
Compute gradient values of the sum-of-squares loss function.
}
