% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{GradientSquaredLossLogIC50}
\alias{GradientSquaredLossLogIC50}
\title{Compute gradient values of the sum-of-squares loss function.}
\usage{
GradientSquaredLossLogIC50(retheta, x, y)
}
\arguments{
\item{retheta}{Parameters among which the IC50 parameter is logarithmically
transformed}

\item{x}{Dose}

\item{y}{Response}
}
\value{
Gradient values of the sum-of-squares loss function.
}
\description{
Compute gradient values of the sum-of-squares loss function.
}
