% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{HessianLogIC50}
\alias{HessianLogIC50}
\title{Compute the Hessian matrix of the sum-of-squares loss function with
reparameterization.}
\usage{
HessianLogIC50(retheta, x, y)
}
\arguments{
\item{retheta}{Parameters of a 4PL model among which the EC50 parameter is
in the log 10 dose scale}

\item{x}{Doses}

\item{y}{Response}
}
\value{
Hessian matrix of the sum-of-squares loss function in terms of
reparameterized parameters.
}
\description{
Compute the Hessian matrix of the sum-of-squares loss function with
reparameterization.
}
