% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{HuberLoss}
\alias{HuberLoss}
\title{Values of Huber's loss function evaluated at residuals r.}
\usage{
HuberLoss(r)
}
\arguments{
\item{r}{Residuals}
}
\value{
result: Huber's loss function values evaluated at residuals r.
}
\description{
Values of Huber's loss function evaluated at residuals r.
}
