% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{Logistic}
\alias{Logistic}
\title{Find initial parameter estimates by the logistic method.}
\usage{
Logistic(x, y, theta.1, theta.4)
}
\arguments{
\item{x}{Vector of dose levels}

\item{y}{Vector of responses}

\item{theta.1}{Estimate of the upper asymptote parameter}

\item{theta.4}{Estimate of the lower asymptote parameter}
}
\value{
Parameter estimates of a 4Pl model in the order from the upper asymptote,
IC50, slope and lower asymptote.
}
\description{
Find initial parameter estimates by the logistic method.
}
