% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{LogToParm}
\alias{LogToParm}
\title{Transform reparameterized parameters (theta.re) back into original parameters 
(theta)}
\usage{
LogToParm(theta.re)
}
\arguments{
\item{theta.re}{Parameters of a 4PL model among which the EC50 parameter is
in the log 10 dose scale}
}
\value{
Parameters of a 4PL model among which the EC50 parameter is in the
dose scale
}
\description{
Transform reparameterized parameters (theta.re) back into original parameters 
(theta)
}
