% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{MeanResponse}
\alias{MeanResponse}
\title{Compute predicted responses.}
\usage{
MeanResponse(x, theta)
}
\arguments{
\item{x}{Dose}

\item{theta}{Parameters}
}
\value{
Predicted response values.
}
\description{
Compute predicted responses.
}

