% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{MeanResponseLogIC50}
\alias{MeanResponseLogIC50}
\title{Compute an estimated mean response with the logarithm IC50 parameter}
\usage{
MeanResponseLogIC50(x, retheta)
}
\arguments{
\item{x}{Dose levels}

\item{retheta}{Parameters among which the IC50 parameter is logarithmically
transformed}
}
\value{
Predicted response values.
}
\description{
Compute an estimated mean response with the logarithm IC50 parameter
}
