% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{ParmToLog}
\alias{ParmToLog}
\title{Transform parameters of a 4PL model (theta) into re-parameterized parameters
(retheta).}
\usage{
ParmToLog(theta)
}
\arguments{
\item{theta}{Parameters of a 4PL model in the dose scale.}
}
\value{
Reparameterized parameters of a 4PL model among which the EC50 parameter
is in the log 10 dose scale.
}
\description{
Transform parameters of a 4PL model (theta) into re-parameterized parameters
(retheta).
}
