% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.DR4PL}
\alias{plot.DR4PL}
\title{plot}
\usage{
\method{plot}{DR4PL}(object, ...)
}
\arguments{
\item{object}{A `DR4PL' object whose mean response function should be plotted.}

\item{...}{All arguments that can normally be passed to plot.}
}
\description{
A default plotting function for a `DR4PL' object.
}
\examples{
ryegrass.DR4PL <- DR4PL::DR4PL(Response ~ Dose, data = sample_data_1)

plot(ryegrass.DR4PL)
}
