% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.drra}
\alias{plot.drra}
\title{plot}
\usage{
\method{plot}{drra}(object, ...)
}
\arguments{
\item{object}{A `drra' object whose mean response function should be plotted.}

\item{...}{All arguments that can normally be passed to plot.}
}
\description{
A default plotting function for a `drra' object.
}
\examples{
ryegrass.drra <- drra::drra(Response ~ Dose, data = sample_data_1)

plot(ryegrass.drra)
}

