% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.DR4PL}
\alias{print.DR4PL}
\title{Print the DR4PL object to screen.}
\usage{
\method{print}{DR4PL}(object, ...)
}
\arguments{
\item{object}{a DR4PL object to be printed}

\item{...}{all normally printable arguments}
}
\description{
Print the DR4PL object to screen.
}
\examples{
ryegrass.DR4PL <- DR4PL(Response ~ Dose,
                      data = sample_data_1)

print(ryegrass.DR4PL)
}
