% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.drra}
\alias{print.drra}
\title{Print the drra object to screen.}
\usage{
\method{print}{drra}(object, ...)
}
\arguments{
\item{object}{a drra object to be printed}

\item{...}{all normally printable arguments}
}
\description{
Print the drra object to screen.
}
\examples{
ryegrass.drra <- drra(Response ~ Dose,
                      data = sample_data_1)

print(ryegrass.drra)
}

