% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print}
\alias{plot.drra}
\alias{print}
\title{print}
\usage{
\method{plot}{drra}(object, ...)
}
\arguments{
\item{object}{A `drra' object whose mean response function should be plotted.
#@examples
ryegrass.drra <- drra::drra(rootl ~ conc, data = drc::ryegrass)

plot(ryegrass.drra)}
}
\description{
A default plotting function for a `drra' object.
}

