% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.summary.DR4PL}
\alias{print.summary.DR4PL}
\title{Print the DR4PL object summary to screen.}
\usage{
\method{print}{summary.DR4PL}(object, ...)
}
\arguments{
\item{object}{a DR4PL object to be summarized}

\item{...}{all normally printable arguments}
}
\description{
Print the DR4PL object summary to screen.
}
