% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.summary.drra}
\alias{print.summary.drra}
\title{Print the drra object summary to screen.}
\usage{
\method{print}{summary.drra}(object, ...)
}
\arguments{
\item{object}{a drra object to be summarized}

\item{...}{all normally printable arguments}
}
\description{
Print the drra object summary to screen.
}

