% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{Residual}
\alias{Residual}
\title{Compute residuals.}
\usage{
Residual(theta, x, y)
}
\arguments{
\item{theta}{Parameters}

\item{x}{Doses}

\item{y}{Responses}
}
\value{
Residuals
}
\description{
Compute residuals.
}
