% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{ResidualLogIC50}
\alias{ResidualLogIC50}
\title{Compute residuals with a reparameterized model.}
\usage{
ResidualLogIC50(retheta, x, y)
}
\arguments{
\item{retheta}{Parameters of a 4pl model among which the IC50 parameter is
transformed into a log 10 scale.}

\item{x}{Vector of doses}

\item{y}{Vector of responses}
}
\value{
Vector of residuals
}
\description{
Compute residuals with a reparameterized model.
}
