% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr4pl-residuals.R
\name{residuals.dr4pl}
\alias{residuals.dr4pl}
\alias{residuals.dr4pl_param}
\title{Compute dr4pl residuals.}
\usage{
\method{residuals}{dr4pl}(object, parm = NULL, ...)

\method{residuals}{dr4pl_param}(object, dose, response, ...)
}
\arguments{
\item{object}{A dr4pl or dr4pl_param object.}

\item{parm}{parameters of the dr4pl object. Usually made with [dr4pl_theta]}

\item{...}{dots for future extensions}

\item{dose}{Dose values}

\item{response}{Response values}
}
\value{
Vector of residuals.
}
\description{
Compute dr4pl residuals.
}
