% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{sample_data_13}
\alias{sample_data_13}
\title{sample_data_13}
\description{
These are a handful of experimentally derived datasets from the wet-laboratory.
These all have numerical errors in other dose-response curve-packages, but not using these methods.
}
\keyword{sample_data}
