% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{sample_data_2}
\alias{sample_data_2}
\title{sample_data_2}
\description{
These are a handful of experimentally derived datasets from the wet-laboratory.
These may or may not have numerical errors in other dose-response curve-packages, but definitly not using these methods.
}
\keyword{sample_data}
