% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary.dr4pl}
\alias{summary.dr4pl}
\title{summary}
\usage{
\method{summary}{dr4pl}(object, ...)
}
\arguments{
\item{object}{a dr4pl object to be summarized}

\item{...}{all normal summary arguments}
}
\description{
Print the dr4pl object summary.
}
