% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary.drra}
\alias{summary.drra}
\title{Print the drra object summary.}
\usage{
\method{summary}{drra}(object, ...)
}
\arguments{
\item{object}{a drra object to be summarized}

\item{...}{all normal summary arguments}
}
\description{
Print the drra object summary.
}

