% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{TukeyBiweightLoss}
\alias{TukeyBiweightLoss}
\title{Values of Tukey's biweight loss function evaluated at residuals r.}
\usage{
TukeyBiweightLoss(r)
}
\arguments{
\item{r}{Residuals}
}
\value{
result: Tukey's biweight loss function values evaluated at residuals r.
}
\description{
Values of Tukey's biweight loss function evaluated at residuals r.
}
