% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{coef.dr4pl}
\alias{coef.dr4pl}
\title{coef}
\usage{
\method{coef}{dr4pl}(object, ...)
}
\arguments{
\item{object}{A 'dr4pl' object}

\item{...}{arguments passed to coef}
}
\value{
A vector of parameters
}
\description{
Coefficient of a `dr4pl' object
}
