% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{DerivativeF}
\alias{DerivativeF}
\title{Compute the Jacobian matrix}
\usage{
DerivativeF(theta, x)
}
\arguments{
\item{theta}{Parameters}

\item{x}{Dose values}
}
\value{
Jacobian matrix
}
\description{
Compute the Jacobian matrix
}
