% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dr4pl.formula}
\alias{dr4pl.formula}
\title{Fit a 4 parameter logistic (4PL) model to dose-response data.}
\usage{
\method{dr4pl}{formula}(formula, constrained = FALSE, data = list(),
  grad = GradientFunction, init.parm = NULL, method.init = "logistic",
  method.optim = if (is.null(grad)) "Nelder-Mead" else "BFGS",
  method.robust = NULL, trace = 0, ...)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit. Either of the
form 'response ~ dose' or as a data frame with response values in first
column and dose values in second column.}

\item{constrained}{Boolean for whether or not this function is constrained.}

\item{data}{A data frame containing variables in the model.}

\item{grad}{The gradient function that returns gradient values of the loss
function specified by \code{method.robust}. If this parameter is NULL, then
the gradient function for the usual sum-of-squares loss functions will be
used.}

\item{init.parm}{A vector of initial parameters to be optimized in the model.}

\item{method.init}{The method of obtaining initial values of the parameters.
If it is NULL, a default "logistic" regression method will be used.}

\item{method.optim}{The method of optimization of the parameters. This method
name is directly applied to the \code{constrOptim} function provided in the
"base" package of R.}

\item{method.robust}{The robust estimation method to be used to fit a model.
- NULL: Sum of squares loss
- absolute: Absolute deviation loss
- Huber: Huber's loss
- Tukey: Tukey's biweight loss}

\item{trace}{Nonnegative number indicating the level of detailedness of output
generated by \code{constrOptim}. As its value increases, more detailed output is
printed.}

\item{...}{Further arguments to be passed to \code{constrOptim}.}
}
\value{
A 'dr4pl' object for which "confint", "gof", "print" and "summary"
  methods are implemented. For details, see the help page of each method.
  For example, type \code{?confint.dr4pl} to obtain the confidence intervals
  of parameters of the 'dr4pl' object.
}
\description{
A general 4PL model fitting function for analysis of
  dose-response relation.
}
\details{
This function fits a 4 parameter logistic (4PL) model to dose-response
  data. A formula of the model is
  \deqn{\theta[1]+(\theta[4]-\theta[1])/(1+(z/\theta[2])^\theta[3])}

  \code{method.init} specifies an initialization method to get initial parameter
  estimates based on data. The currently supported initialization methods are
  'logistic' and 'Mead'. Details of the methods are given in Section
  'Details'.

  \code{method.optim} specifies an optimization method to be used in
  "constrOptim" function. The currently supported optimization techniques
  include "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN" and "Brent". For
  further details, see the help page of \code{\link[stats]{optim}}.

  \code{method.robust} chooses a robust estimation method among 4 methods.
  The method of estimation is usually identified by the loss function of the
  method. This package implements 4 loss functions: sum of squares loss,
  absolute deviation loss, Huber's loss and Tukey's biweight loss. Each of
  loss function is explained in detail in the vignette.
}
\examples{
ryegrass.dr4pl <- dr4pl(Response ~ Dose, data = sample_data_1)

ryegrass.dr4pl
}
\seealso{
\code{\link{confint.dr4pl}}, \code{\link{gof.dr4pl}},
\code{\link{print.dr4pl}}, \code{\link{summary.dr4pl}}
}
\author{
Hyowon An, Dirk P. Dittmer and J. S. Marron
}
