% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{dr4pl}
\alias{dr4pl}
\alias{coef.dr4pl}
\alias{dr4pl-package}
\alias{dr4pl}
\alias{dr4pl.default}
\title{Dose response data analysis using the 4 Parameter Logistic model}
\usage{
\method{coef}{dr4pl}(object, ...)

dr4pl(x, ...)

\method{dr4pl}{default}(dose, response, constrained = FALSE,
  grad = GradientFunction, init.parm = NULL, method.init = "logistic",
  method.optim = if (is.null(grad)) "Nelder-Mead" else "BFGS",
  method.robust = NULL, trace = 0, ...)
}
\arguments{
\item{object}{A 'dr4pl' object}

\item{...}{arguments passed to coef}

\item{x}{Object to dr4pl.}

\item{dose}{Dose}

\item{response}{Response}

\item{...}{arguments passed to coef}
}
\value{
A vector of parameters
}
\description{
Main functions related to fitting

Coefficient of a `dr4pl' object

Fit the 4 parameter logistic model to the data using the function `dr4plEst'
}
\details{
Last updated: 09/19/2016
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Used in the default case, supplying a single dose and response variable
}}

\author{
Hyowon An, Lineberger Comprehensive Cancer Center
}
