% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{dr4pl}
\alias{dr4pl}
\alias{confint.dr4pl}
\alias{dr4pl-package}
\alias{dr4pl}
\alias{dr4pl.default}
\title{Dose response data analysis using the 4 Parameter Logistic model}
\usage{
\method{confint}{dr4pl}(object, parm, level = 0.05, ...)

dr4pl(...)

\method{dr4pl}{default}(dose, response, grad = GradientSquaredLoss,
  init.parm = NULL, method.init = "logistic", method.optim = if
  (is.null(grad)) "Nelder-Mead" else "BFGS", method.robust = NULL, ...)
}
\arguments{
\item{object}{An object of the dr4pl class.}

\item{parm}{parameter of the dr4pl class. This argument may be 
ignored when object is assigned a dr4pl class.}

\item{level}{Sigifigance level of the confidence intervals}

\item{...}{additional argument(s) for methods}

\item{dose}{Dose}

\item{response}{Response}

\item{grad}{The gradient function that returns gradient values of the loss
function specified by \code{method.robust}. If this parameter is NULL, then
the gradient function for the usual sum-of-squares loss functions will be
used.}

\item{init.parm}{A vector of initial parameters to be optimized in the model.}

\item{method.init}{The method of obtaining initial values of the parameters.
If it is NULL, a default "logistic" regression method will be used. Assign
"Mead" to use Mead's method.}

\item{method.optim}{The method of optimization of the parameters. This method
name is directly applied to the \code{constrOptim} function provided in the
"base" package of R.}

\item{method.robust}{Parameter to select error function for the robust estimation method to be used to fit a model. 
- NULL: Sum of squares loss 
- absolute: Absolute deviation loss 
- Huber: Huber's loss 
- Tukey: Tukey's biweight loss}

\item{...}{Dose Response dataframe to dr4pl object. Use either formula or direct argument assignment}
}
\value{
A matrix of the confidence intervals in which each row represents a
  parameter and each column represents the lower and upper bounds of the
  confidence intervals of the corresponding parameters.
}
\description{
Main functions related to fitting

Compute the confidence intervals of parameter estimates of a fitted
  model.

Fit the 4 parameter logistic model to the data using the function `dr4plEst'
}
\details{
Last updated: 09/19/2016

This function computes the confidence intervals of the parameters of the
  4PL model based on the second order approximation to the Hessian matrix of the
  loss function of the model. Refer to Subsection 5.2.2 of 
  Seber, G. A. F. and Wild, C. J. (1989). Nonlinear Regression. Wiley Series in
  Probability and Mathematical Statistics: Probability and Mathematical
  Statistics. John Wiley & Sons, Inc., New York.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Used in the default case, supplying a single dose and 
response variable
}}

\examples{
  obj.dr4pl <- dr4pl(Response ~ Dose, data = sample_data_1)

  confint(obj.dr4pl)

  a <- dr4pl(dose = sample_data_1$Dose, 
               response = sample_data_1$Response, 
               method.init = "logistic")
  plot(a)

  ##Assign method.init = "Mead" to use Mead's method of estimation. 
  # Use method.robust to select desired error function
  b <- dr4pl(Response~Dose, 
               data = sample_data_1,
               method.init = "Mead", 
               method.robust = "Tukey" )
  plot(b)

  ##compatable with ggplot
  library(ggplot2)
  c <- dr4pl(Response~Dose, 
             data = drc_error_2,
             method.init = "Mead", 
             method.robust = "absolute" )
  d <- plot(c)
  d + scale_x_log10(breaks = c(.00135, .0135, .135, 1.35, 13.5))
}
\author{
Hyowon An, Lineberger Comprehensive Cancer Center

Hyowon An, Justin T. Landis and Aubrey G. Bailey
}
