% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{dr4pl}
\alias{dr4pl}
\alias{dr4plEst}
\alias{dr4pl-package}
\alias{dr4pl}
\alias{dr4pl.default}
\title{Dose response data analysis using the 4 Parameter Logistic model}
\usage{
dr4plEst(dose, response, constrained = constrained, grad, init.parm,
  method.init, method.optim, method.robust, trace = 0)

dr4pl(x, ...)

\method{dr4pl}{default}(dose, response, constrained = FALSE, grad = NULL,
  init.parm = NULL, method.init = "logistic", method.optim = if
  (is.null(grad)) "Nelder-Mead" else "BFGS", method.robust = NULL,
  trace = 0, ...)
}
\arguments{
\item{dose}{Dose}

\item{response}{Response}

\item{constrained}{Boolean for whether or not this function is constrained.}

\item{grad}{Gradient function}

\item{init.parm}{Vector of initial parameters}

\item{method.init}{Initialization method}

\item{method.optim}{Optimization method}

\item{method.robust}{Robust estimation method
- NULL: Sum of squares loss
- absolute: Absolute deviation loss
- Huber: Huber's loss
- Tukey: Tukey's biweight loss}

\item{trace}{Nonnegative number indicating the level of detailedness of output
generated by \code{optim}. As its value increases, more detailed output is
printed.}

\item{x}{Object to dr4pl.}

\item{...}{arguments passed to coef}
}
\description{
Main functions related to fitting

Fit the 4 parameter logistic model to the data using the function `dr4plEst'
}
\details{
Last updated: 09/19/2016
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Used in the default case, supplying a single dose and response variable
}}

\author{
Hyowon An, Lineberger Comprehensive Cancer Center
}
