% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{dr4pl}
\alias{dr4pl}
\alias{coef.dr4pl}
\alias{dr4pl-package}
\alias{dr4pl}
\alias{dr4pl.default}
\title{Dose response data analysis using the 4 Parameter Logistic model}
\usage{
\method{coef}{dr4pl}(object, ...)

dr4pl(...)

\method{dr4pl}{default}(dose, response, constrained = FALSE,
  grad = GradientSquaredLoss, init.parm = NULL, method.init = "logistic",
  method.optim = if (is.null(grad)) "Nelder-Mead" else "BFGS",
  method.robust = NULL, ...)
}
\arguments{
\item{object}{A 'dr4pl' object}

\item{...}{arguments passed to coef}

\item{dose}{Dose}

\item{response}{Response}

\item{constrained}{Boolean for whether or not this function is constrained.}

\item{grad}{The gradient function that returns gradient values of the loss
function specified by \code{method.robust}. If this parameter is NULL, then
the gradient function for the usual sum-of-squares loss functions will be
used.}

\item{init.parm}{A vector of initial parameters to be optimized in the model.}

\item{method.init}{The method of obtaining initial values of the parameters.
If it is NULL, a default "logistic" regression method will be used. Assign
"Mead" to use Mead's method.}

\item{method.optim}{The method of optimization of the parameters. This method
name is directly applied to the \code{constrOptim} function provided in the
"base" package of R.}

\item{method.robust}{The robust estimation method to be used to fit a model.
- NULL: Sum of squares loss
- absolute: Absolute deviation loss
- Huber: Huber's loss
- Tukey: Tukey's biweight loss}

\item{...}{Dose Response dataframe to dr4pl object. Use either formula or direct argument assignment}
}
\value{
A vector of parameters
}
\description{
Main functions related to fitting

Coefficient of a `dr4pl' object

Fit the 4 parameter logistic model to the data using the function `dr4plEst'
}
\details{
Last updated: 09/19/2016
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Used in the default case, supplying a single dose and 
response variable
}}

\author{
Hyowon An, Lineberger Comprehensive Cancer Center
}
