% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{dr4pl}
\alias{dr4pl}
\alias{dr4pl-package}
\alias{dr4pl.default}
\title{Dose response data analysis using the 4 Parameter Logistic model}
\usage{
dr4pl(...)

\method{dr4pl}{default}(dose, response, grad = GradientSquaredLoss,
  init.parm = NULL, method.init = "logistic", method.optim = if
  (is.null(grad)) "Nelder-Mead" else "BFGS", method.robust = NULL, ...)
}
\arguments{
\item{...}{Dose Response dataframe to dr4pl object. Use either formula or direct argument assignment}

\item{dose}{Dose}

\item{response}{Response}

\item{grad}{The gradient function that returns gradient values of the loss
function specified by \code{method.robust}. If this parameter is NULL, then
the gradient function for the usual sum-of-squares loss functions will be
used.}

\item{init.parm}{A vector of initial parameters to be optimized in the model.}

\item{method.init}{The method of obtaining initial values of the parameters.
If it is NULL, a default "logistic" regression method will be used. Assign
"Mead" to use Mead's method.}

\item{method.optim}{The method of optimization of the parameters. This method
name is directly applied to the \code{constrOptim} function provided in the
"base" package of R.}

\item{method.robust}{Parameter to select error function for the robust estimation method to be used to fit a model. 
- NULL: Sum of squares loss 
- absolute: Absolute deviation loss 
- Huber: Huber's loss 
- Tukey: Tukey's biweight loss}
}
\description{
Main functions related to fitting

Fit the 4 parameter logistic model to the data using the function `dr4plEst'
}
\details{
Last updated: 09/19/2016
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Used in the default case, supplying a single dose and 
response variable
}}

\examples{
  a <- dr4pl(dose = sample_data_1$Dose, 
               response = sample_data_1$Response, 
               method.init = "logistic")
  plot(a)

  ##Assign method.init = "Mead" to use Mead's method of estimation. 
  # Use method.robust to select desired error function
  b <- dr4pl(Response~Dose, 
               data = sample_data_1,
               method.init = "Mead", 
               method.robust = "Tukey" )
  plot(b)

  ##compatable with ggplot
  library(ggplot2)
  c <- dr4pl(Response~Dose, 
             data = drc_error_2,
             method.init = "Mead", 
             method.robust = "absolute" )
  d <- plot(c)
  d + scale_x_log10(breaks = c(.00135, .0135, .135, 1.35, 13.5))
}
\author{
Hyowon An, Lineberger Comprehensive Cancer Center
}
