% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{drc_error_1}
\alias{drc_error_1}
\title{Single High Outlier}
\description{
These are a handful of experimentally derived datasets from the wet-laboratory.
These all have numerical errors in other dose-response curve-packages, but not using these methods.
This data set exemplifies the case of a single extreme outlier of one dose measurement.
}
\examples{
  a <- dr4pl(Response~Dose, data = drc_error_1, method.init = "logistic", method.robust = "Tukey")
  plot(a)
}
\keyword{sample_data}
