% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{drc_error_3}
\alias{drc_error_3}
\title{Support Problem and Outliers at a Single Dose Level}
\description{
These are a handful of experimentally derived datasets from the wet-laboratory.
These all have numerical errors in other dose-response curve-packages, but not using these methods.
}
\keyword{sample_data}
