% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{FindInitialParms}
\alias{FindInitialParms}
\title{Find initial values for the 4PL model.}
\usage{
FindInitialParms(x, y, method.init, method.robust)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{method.init}{Initialization method}

\item{method.robust}{Robust fitting method}
}
\value{
theta.IC50.slope Parameter estimates of the IC50 and slope
}
\description{
Args:
}
