% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{FindInitialParms}
\alias{FindInitialParms}
\title{Find initial values for the 4PL model.}
\usage{
FindInitialParms(x, y, decline, method.init, method.robust)
}
\arguments{
\item{x}{Vector of dose levels}

\item{y}{Vector of responses}

\item{decline}{Indicator of whether the curve is a decline \eqn{\theta[3]<0} 
or growth curve \eqn{\theta[3]>0}. See \code{\link{dr4pl}} for detailed 
explanation.}

\item{method.init}{Method of obtaining initial values of the parameters. See
\code{\link{dr4pl}} for detailed explanation.}

\item{method.robust}{Parameter to select loss function for the robust estimation 
method to be used to fit a model. See \code{\link{dr4pl}} for detailed
explanation.}
}
\value{
Initial parameter estimates of a 4PL model in the order of
}
\description{
Find initial values for the 4PL model.
}
