% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{FindInitialParms}
\alias{FindInitialParms}
\title{FindInitialParms}
\usage{
FindInitialParms(x, y, trend = "auto", method.init = "Mead",
  method.robust = "squared")
}
\arguments{
\item{x}{Vector of dose levels}

\item{y}{Vector of responses}

\item{trend}{Indicator of whether the curve is a decreasing \eqn{\theta[3]<0} 
or increasing curve \eqn{\theta[3]>0}. See \code{\link{dr4pl}} for detailed 
explanation.}

\item{method.init}{Method of obtaining initial values of the parameters. See
\code{\link{dr4pl}} for detailed explanation.}

\item{method.robust}{Parameter to select loss function for the robust estimation 
method to be used to fit a model. See \code{\link{dr4pl}} for detailed
explanation.}
}
\value{
Initial parameter estimates of a 4PL model in the order of the
upper asymptote, IC50, Slope and lower asymptote parameters.
}
\description{
Find initial parameter estimates for a 4PL model.
}
