% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{GradientFunction}
\alias{GradientFunction}
\title{Compute gradient values.}
\usage{
GradientFunction(theta, dose, response)
}
\arguments{
\item{theta}{Parameters}

\item{dose}{Dose}

\item{response}{Response}
}
\value{
Gradient values.
}
\description{
Compute gradient values.
}

