% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{Hessian}
\alias{Hessian}
\title{Compute the Hessian matrix of the sum-of-squares loss function}
\usage{
Hessian(theta, x, y)
}
\arguments{
\item{theta}{Parameters}

\item{x}{Doses}

\item{y}{Response}
}
\value{
A hessian matrix
}
\description{
Compute the Hessian matrix of the sum-of-squares loss function
}
