% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{MeanResponse}
\alias{MeanResponse}
\title{Compute an estimated mean response}
\usage{
MeanResponse(x, theta)
}
\arguments{
\item{x}{Dose levels}

\item{theta}{Parameters of the 4PL model}
}
\value{
Predicted response values.
}
\description{
Compute an estimated mean response
}
