% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr4pl-MeanResponse.R
\name{MeanResponse}
\alias{MeanResponse}
\alias{MeanResponse.dr4pl}
\alias{MeanResponse.numeric}
\alias{MeanResponse.dr4pl_theta}
\alias{MeanResponse.dr4pl_log10}
\title{Compute an estimated mean response.}
\usage{
MeanResponse(...)

\method{MeanResponse}{dr4pl}(dr4pl, theta = NULL, ...)

\method{MeanResponse}{numeric}(theta, x, ...)

\method{MeanResponse}{dr4pl_theta}(theta, x, ...)

\method{MeanResponse}{dr4pl_log10}(theta, x, ...)
}
\arguments{
\item{...}{arguments to be passed to S3 methods}

\item{dr4pl}{dr4pl object}

\item{theta}{Parameters of the dr4pl object. Usually made with [dr4pl_theta]}

\item{x}{domain values for 4PL model. Values
should always be passed to this function on the
linear space.}
}
\value{
Predicted response values.
}
\description{
Compute an estimated mean response.
}
