% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.dr4pl}
\alias{plot.dr4pl}
\title{plot}
\usage{
\method{plot}{dr4pl}(x, indices.outlier = NULL,
  text.title = "Dose response plot", ...)
}
\arguments{
\item{x}{`dr4pl' object whose mean response function should be plotted.}

\item{indices.outlier}{Indices of outliers in data.}

\item{text.title}{Character string for the title of a plot.}

\item{...}{All arguments that can normally be passed to ggplot.}
}
\description{
Default plotting function for a `dr4pl' object.
}
\examples{
ryegrass.dr4pl <- dr4pl::dr4pl(Response ~ Dose, data = sample_data_1)

plot(ryegrass.dr4pl)
}
