% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.dr4pl}
\alias{plot.dr4pl}
\title{plot}
\usage{
\method{plot}{dr4pl}(x, text.title = "Dose response plot",
  indices.outlier = NULL, ...)
}
\arguments{
\item{x}{`dr4pl' object whose mean response function should be plotted.}

\item{text.title}{Character string for the title of a plot}

\item{indices.outlier}{Indices of outliers in data}

\item{...}{All arguments that can normally be passed to ggplot.}
}
\description{
Default plotting function for a `dr4pl' object. Plot displays 
decreasing dr4pl curve as well as measured points. Default points are 
blue and size 5.
}
\examples{
ryegrass.dr4pl <- dr4pl::dr4pl(Response ~ Dose, data = sample_data_1)

plot(ryegrass.dr4pl)

#Able to further edit plots
library(ggplot2)
ryegrass.dr4pl <- dr4pl::dr4pl(Response ~ Dose, data = sample_data_1, text.title = "Sample Data Plot")

a <- plot(ryegrass.dr4pl) 
a + geom_point(color = "green", size = 5)
}
