% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.dr4pl}
\alias{plot.dr4pl}
\title{plot}
\usage{
\method{plot}{dr4pl}(object, text.title = "Dose response plot",
  indices.outlier = NULL, ...)
}
\arguments{
\item{...}{All arguments that can normally be passed to plot.}

\item{x}{A `dr4pl' object whose mean response function should be plotted.}
}
\description{
A default plotting function for a `dr4pl' object.
}
\examples{
ryegrass.dr4pl <- dr4pl::dr4pl(Response ~ Dose, data = sample_data_1)

plot(ryegrass.dr4pl)
}
