% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{print.dr4pl}
\alias{print.dr4pl}
\title{Print the dr4pl object to screen.}
\usage{
\method{print}{dr4pl}(object, ...)
}
\arguments{
\item{object}{a dr4pl object to be printed}

\item{...}{all normally printable arguments}
}
\description{
Print the dr4pl object to screen.
}
\examples{
ryegrass.dr4pl <- dr4pl(Response ~ Dose,
                      data = sample_data_1)

print(ryegrass.dr4pl)
}
