% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.drra}
\alias{print.drra}
\title{Print the drra object to screen.}
\usage{
\method{print}{drra}(x, ...)
}
\arguments{
\item{x}{A drra object.}
}
\description{
Print the drra object to screen.
}
\examples{
ryegrass.drra <- drra(rootl ~ conc,
                      data = drc::ryegrass)

print(ryegrass.drra)
}

