% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.summary.dr4pl}
\alias{print.summary.dr4pl}
\title{Print the dr4pl object summary to screen.}
\usage{
\method{print}{summary.dr4pl}(object, ...)
}
\arguments{
\item{object}{a dr4pl object to be summarized}

\item{...}{all normally printable arguments}
}
\description{
Print the dr4pl object summary to screen.
}
