% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{print.summary.dr4pl}
\alias{print.summary.dr4pl}
\title{Print the dr4pl object summary to screen.}
\usage{
\method{print}{summary.dr4pl}(object, ...)
}
\arguments{
\item{object}{a dr4pl object to be summarized}

\item{...}{all normally printable arguments}
}
\description{
Print the dr4pl object summary to screen.
}
\examples{
library(drc) #needed for ryegrass data set
ryegrass.dr4pl <- dr4pl(rootl ~ conc, data = ryegrass)
print(summary(ryegrass.dr4pl))

obj.dr4pl <- dr4pl(Response ~ Dose, data = sample_data_7)
print(summary(obj.dr4pl))
}
