% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{Residual}
\alias{Residual}
\title{Compute residuals.}
\usage{
Residual(theta, x, y)
}
\arguments{
\item{theta}{Parameters of a 4PL model}

\item{x}{Vector of doses}

\item{y}{Vector of responses}
}
\value{
Vector of residuals
}
\description{
Compute residuals.
}
