% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary.DR4PL}
\alias{summary.DR4PL}
\title{Print the DR4PL object summary.}
\usage{
\method{summary}{DR4PL}(object, ...)
}
\arguments{
\item{object}{a DR4PL object to be summarized}

\item{...}{all normal summary arguments}
}
\description{
Print the DR4PL object summary.
}
