% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{summary.dr4pl}
\alias{summary.dr4pl}
\title{summary}
\usage{
\method{summary}{dr4pl}(object, ...)
}
\arguments{
\item{object}{a dr4pl object to be summarized}

\item{...}{all normal summary arguments}
}
\description{
Print the summary of a dr4pl object.
}
\examples{
obj.dr4pl <- dr4pl(Response ~ Dose, data = sample_data_5)  # Fit a 4PL model to data
summary(obj.dr4pl)

obj.dr4pl <- dr4pl(Response ~ Dose, data = sample_data_6)  # Fit a 4PL model to data
summary(obj.dr4pl)

}
